requirements.txt
httpx
loguru
python-telegram-bot[job-queue]
PyYAML
